+++ src/app/grid-search/grid-search.sample.ts
@@ -26,19 +26,19 @@ export class GridSearchComponent implements OnInit {
         ];
 
         this.columns = [
-            { field: 'ID', width: 80, resizable: true, movable: true },
-            { field: 'CompanyName', width: 150, resizable: true, movable: true, type: 'string'},
-            { field: 'ContactName', width: 150, resizable: true, movable: true, type: 'string' },
-            { field: 'Employees', width: 150, resizable: true, movable: true, type: 'number' },
-            { field: 'ContactTitle', width: 150, resizable: true, movable: true, type: 'string' },
-            { field: 'DateCreated', width: 150, resizable: true, movable: true, type: 'date' },
-            { field: 'Address', width: 150, resizable: true, movable: true, type: 'string' },
-            { field: 'City', width: 150, resizable: true, movable: true, type: 'string' },
-            { field: 'Region', width: 150, resizable: true, movable: true, type: 'string' },
-            { field: 'PostalCode', width: 150, resizable: true, movable: true, type: 'string' },
-            { field: 'Phone', width: 150, resizable: true, movable: true, type: 'string' },
-            { field: 'Fax', width: 150, resizable: true, movable: true, type: 'string' },
-            { field: 'Contract', width: 150, resizable: true, movable: true, type: 'boolean' }
+            { field: 'ID', width: 80, resizable: true },
+            { field: 'CompanyName', width: 150, resizable: true, type: 'string'},
+            { field: 'ContactName', width: 150, resizable: true, type: 'string' },
+            { field: 'Employees', width: 150, resizable: true, type: 'number' },
+            { field: 'ContactTitle', width: 150, resizable: true, type: 'string' },
+            { field: 'DateCreated', width: 150, resizable: true, type: 'date' },
+            { field: 'Address', width: 150, resizable: true, type: 'string' },
+            { field: 'City', width: 150, resizable: true, type: 'string' },
+            { field: 'Region', width: 150, resizable: true, type: 'string' },
+            { field: 'PostalCode', width: 150, resizable: true, type: 'string' },
+            { field: 'Phone', width: 150, resizable: true, type: 'string' },
+            { field: 'Fax', width: 150, resizable: true, type: 'string' },
+            { field: 'Contract', width: 150, resizable: true, type: 'boolean' }
         ];
         this.data = [
             {